# cmd/api/docs/swagger.yaml
definitions:
  entity.Match:
    properties:
      away_team_id:
        type: integer
      created_at:
        type: string
      deleted_at:
        description: DeletedAt *gorm.DeletedAt `gorm:"index" json:"deleted_at"`
        type: string
      home_team_id:
        type: integer
      id:
        type: integer
      match_date:
        type: string
      match_time:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  entity.Team:
    properties:
      address:
        type: string
      city:
        type: string
      created_at:
        type: string
      deleted_at:
        description: DeletedAt *gorm.DeletedAt `gorm:"index" json:"deleted_at"`
        type: string
      founded_year:
        type: integer
      id:
        type: integer
      logo_url:
        type: string
      name:
        type: string
      total_wins:
        type: integer
      updated_at:
        type: string
    type: object
...
  request.GoalRequest:
    properties:
      is_own_goal:
        type: boolean
      minute:
        type: integer
      player_id:
        type: integer
    required:
    - minute
    - player_id
    type: object
...
  request.ReportMatchRequest:
    properties:
      away_score:
        type: integer
      goals:
        items:
          $ref: '#/definitions/request.GoalRequest'
        type: array
      home_score:
        type: integer
    required:
    - away_score
    - goals
    - home_score
    type: object
  request.ScheduleMatchRequest:
    properties:
      away_team_id:
        type: integer
      home_team_id:
        type: integer
      match_date:
        description: 'format: "2026-01-02"'
        type: string
      match_time:
        description: 'format: "15:04:05"'
        type: string
    required:
    - away_team_id
    - home_team_id
    - match_date
    - match_time
    type: object
  response.BaseResponse:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
...
  response.MatchGoalData:
    properties:
      id:
        type: integer
      is_own_goal:
        type: boolean
      minute:
        type: integer
      player_id:
        type: integer
      player_name:
        type: string
    type: object
  response.MatchReportData:
    properties:
      away_score:
        type: integer
      away_team:
        $ref: '#/definitions/entity.Team'
      away_total_wins:
        type: integer
      goals:
        items:
          $ref: '#/definitions/response.MatchGoalData'
        type: array
      home_score:
        type: integer
      home_team:
        $ref: '#/definitions/entity.Team'
      home_total_wins:
        type: integer
      match:
        $ref: '#/definitions/entity.Match'
      top_scorer:
        type: string
      top_scorer_goal_count:
        type: integer
      winner:
        type: string
    type: object
  response.MatchReportResponse:
    properties:
      data:
        $ref: '#/definitions/response.MatchReportData'
      message:
        type: string
      status:
        type: integer
    type: object
  response.MatchResponse:
    properties:
      match:
        $ref: '#/definitions/entity.Match'
      message:
        type: string
      status:
        type: integer
    type: object
  response.MatchesResponse:
    properties:
      matches:
        items:
          $ref: '#/definitions/entity.Match'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMetadata'
      status:
        type: integer
    type: object
  response.PaginationMetadata:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_data:
        type: integer
      total_page:
        type: integer
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.swagger.io/support
  description: This is the documentation for the main e-commerce service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Ayo Test API
  version: "1.0"
paths:
...
  /matches:
    get:
      consumes:
      - application/json
      description: Get all matches with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MatchesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: List Matches
      tags:
      - matches
    post:
      consumes:
      - application/json
      description: Create a new match schedule (admin only)
      parameters:
      - description: Match schedule data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ScheduleMatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MatchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Schedule Match
      tags:
      - matches
  /matches/{id}:
    delete:
      description: Delete match by ID (admin only, soft delete)
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Match
      tags:
      - matches
    get:
      description: Get match by ID
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MatchResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Match
      tags:
      - matches
    put:
      consumes:
      - application/json
      description: Update match schedule by ID (admin only)
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: integer
      - description: Match schedule data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ScheduleMatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MatchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Match
      tags:
      - matches
  /matches/{id}/report:
    get:
      description: Get match result with goals, winner, top scorer, and team win accumulations
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MatchReportResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Match Report
      tags:
      - matches
    post:
      consumes:
      - application/json
      description: Submit match result with goals (admin only)
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: integer
      - description: Match result data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ReportMatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Report Match Result
      tags:
      - matches
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"